fix(process): widen Register::X from u8 to u16 to match X register array, bump v0.3.6
The Register::X enum variant and messaging.rs register() were still
using u8 after the X register array was widened to 1024 in v0.3.5.
This caused InvalidOperand("X register") errors when OTP 27 code used
register indices > 255 in try/catch or receive contexts.

+++ crates/beamr/src/process/mod.rs
@@ -77,7 +77,7 @@ pub struct CodePosition {
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 pub enum Register {
     /// X register index.
-    X(u8),
+    X(u16),
     /// Y register index in the current stack frame.
     Y(u16),
 }
